visuals

+++ src/webview/breadboard/ComponentDropdown.svelte
@@ -2,7 +2,7 @@
     import { createEventDispatcher } from 'svelte'
     const dispatch = createEventDispatcher()
     import DetailDropdown from './controls/DetailDropdown'
-    import iconDropdown from '@vscode/codicons/src/icons/tools.svg'
+    import iconDropdown from '@vscode/codicons/src/icons/ellipsis.svg'
 
     import { extension } from './transport'
 
@@ -18,15 +18,19 @@
             viewColumn: 'Beside',
         })
     }
+
+    let canDelete = true
 </script>
 
 <details use:DetailDropdown class="dropdown select-none dropdown--having-svg-in-summary">
     <summary class="cursor-pointer">{@html iconDropdown}</summary>
     <div class="menu menu--vertical-padding widget-shadow" style="max-width:300px">
-        <div class="menu__item" on:click="{()=>dispatch('rename')}"><span class="grow truncate">Rename component</span></div>
-        <div class="menu__item" on:click="{()=>dispatch('delete')}"><span class="grow truncate">Delete component</span></div>
+        <div class="menu__item" on:click="{()=>dispatch('rename')}"><span class="grow truncate">Rename</span></div>
         {#if canReveal}
-            <div class="menu__item" on:click="{()=>reveal()}"><span class="grow truncate">Reveal component</span></div>
+            <div class="menu__item" on:click="{()=>reveal()}"><span class="grow truncate">Reveal code</span></div>
+        {/if}
+        {#if canDelete}
+            <div class="menu__item" on:click="{()=>dispatch('delete')}"><span class="grow truncate">Delete component</span></div>
         {/if}
     </div>
 </details>
